requirements.txt
django==2.2.*
djangorestframework==3.11.*
django-environ==0.4.*
psycopg2-binary==2.8.*
Pillow==7.1.*
factory-boy==2.12.*
